+++ app/(auth)/signup/page.tsx
@@ -16,7 +16,7 @@ import { getFieldMessage } from '@/app/shared/consts/errorMessages';
 import { FormStatus } from '@/app/shared/types/FormStatus';
 import { Button } from '@/app/shared/ui/button';
 import Input from '@/app/shared/ui/input';
-import Label from '@/app/shared/ui/label/Label';
+import Label from '@/app/shared/ui/label';
 import { useRouter } from 'next/navigation';
 
 export default function SignupPage() {
